+++ BreakoutGame/src/domain/Classroom.java
@@ -31,13 +31,13 @@ public class Classroom implements IManageable {
     }
       
     @Override
-    public long getId()
+    public int getId()
     {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
 
     @Override
-    public void setId(long id)
+    public void setId(int id)
     {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
