+++ BreakoutGame/src/domain/managers/IManageable.java
@@ -9,7 +9,7 @@ package domain.managers;
  */
 public interface IManageable
 {
-    long getId();
-    void setId(long id);
+    int getId();
+    void setId(int id);
     
 }
